Add UNIQUE constraint to tokens.token_hash
We're never going to ever hit a sha256 collision, but let's go ahead and
record that yes, we are in fact hard-assuming this. And if we do, we get a clear
error instead of a ghost infestation.

Plus, token_hash is the key we always look up by, and if I understand correctly,
a UNIQUE constraint automatically creates an index on that column for a nice
speed-up.