Omit Dogear.display_name in Serialize if None
I dislike this and haven't fully evaluated what effect it's gonna have on my
json interfaces, but minijinja does something _really_ weird with Option::None
values and I'm gonna need to do something to work around it. This does the trick
for now.

I was shocked enough to file a bug for this.
https://github.com/mitsuhiko/minijinja/issues/444