Add `short_date` filter to template environment, use it on dogear.updated
Eardogger 1 uses mm/dd/yyyy for its human readable shortdates, which I think is
an emergent property of whatever computer happens to be running the Node VM,
because I'm just using .toLocaleDateString() with no args, lolll.

Anyway, I'm gonna switch to yyyy-mm-dd, because it's more logical and
world-friendly.

Round-tripping that OffsetDateTime through an iso8601 string and back feels
pointlessly wasteful... BUT, the alternative is to write VASTLY messier
serialization code or template value glue code. And I hate that more!!!