Add POST /changepassword route
I'm removing the change_password DB method, because the frontend needs to be
able to return 400 bad request for password validation problems, and I've come
around to the convention that an error out of a DB handler means a 500 internal
server error. So, we'll just put the new password validation checks into a
new util method to make em easier to remember.