Query helpers: Wrap ref to full `Db` struct, use appropriate pools
Tests all still pass!

This adds a bit of boilerplate to tersely access the pools. The reason I'm
wrapping the whole Db struct instead of the two pools is, it's also going to
acquire a TaskTracker for spawning fire-and-forget updates off the critical
return path, and the query helpers will want access to that as well.