Manually configure tokio runtime and max DB readers
That's it for the `#[tokio::main]` macro. See u, space cowboy 💋

It turns out the multithreaded runtime defaults to using *the number of logical
cores* as its thread count. This is somewhat antisocial! And I have a suspicion
it's at the root of some new problems I started having on my webhost after they
pushed an OS update.